introduced ProductFilterDataFactory in the framework
