+++ packages/framework/src/Model/Product/Filter/ProductFilterDataFactory.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopsys\FrameworkBundle\Model\Product\Filter;
+
+class ProductFilterDataFactory
+{
+    /**
+     * @return \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterData
+     */
+    public function create(): ProductFilterData
+    {
+        return new ProductFilterData();
+    }
+}
