+++ project-base/app/src/FrontendApi/Model/Product/Filter/ProductFilterDataMapper.php
@@ -5,11 +5,6 @@ declare(strict_types=1);
 namespace App\FrontendApi\Model\Product\Filter;
 
 use App\Model\Product\Filter\ParameterFilterData;
-use App\Model\Product\Filter\ProductFilterDataFactory;
-use Shopsys\FrameworkBundle\Model\Product\Brand\BrandFacade;
-use Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterData;
-use Shopsys\FrameworkBundle\Model\Product\Flag\FlagFacade;
-use Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterFacade;
 use Shopsys\FrontendApiBundle\Model\Product\Filter\ProductFilterDataMapper as BaseProductFilterDataMapper;
 
 /**
@@ -18,24 +13,12 @@ use Shopsys\FrontendApiBundle\Model\Product\Filter\ProductFilterDataMapper as Ba
  * @property \App\Model\Product\Parameter\ParameterFacade $parameterFacade
  * @property \App\Model\Product\Parameter\Parameter[] $parametersByUuid
  * @property \App\Model\Product\Parameter\ParameterValue[] $parameterValuesByUuid
+ * @method \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterData mapFrontendApiFilterToProductFilterData(array $frontendApiFilter)
+ * @property \App\Model\Product\Filter\ProductFilterDataFactory $productFilterDataFactory
+ * @method __construct(\App\Model\Product\Flag\FlagFacade $flagFacade, \App\Model\Product\Brand\BrandFacade $brandFacade, \App\Model\Product\Parameter\ParameterFacade $parameterFacade, \App\Model\Product\Filter\ProductFilterDataFactory $productFilterDataFactory)
  */
 class ProductFilterDataMapper extends BaseProductFilterDataMapper
 {
-    /**
-     * @param \App\Model\Product\Flag\FlagFacade $flagFacade
-     * @param \App\Model\Product\Brand\BrandFacade $brandFacade
-     * @param \App\Model\Product\Parameter\ParameterFacade $parameterFacade
-     * @param \App\Model\Product\Filter\ProductFilterDataFactory $productFilterDataFactory
-     */
-    public function __construct(
-        FlagFacade $flagFacade,
-        BrandFacade $brandFacade,
-        ParameterFacade $parameterFacade,
-        private ProductFilterDataFactory $productFilterDataFactory,
-    ) {
-        parent::__construct($flagFacade, $brandFacade, $parameterFacade);
-    }
-
     /**
      * @param array $parameterAndValueUuids
      * @return \App\Model\Product\Filter\ParameterFilterData[]
@@ -77,32 +60,5 @@ class ProductFilterDataMapper extends BaseProductFilterDataMapper
         }
 
         return $parametersFilterData;
-    }
-
-    /**
-     * Method is extended because of https://github.com/shopsys/shopsys/pull/2380
-     *
-     * @param array $frontendApiFilter
-     * @return \App\Model\Product\Filter\ProductFilterData
-     */
-    public function mapFrontendApiFilterToProductFilterData(array $frontendApiFilter): ProductFilterData
-    {
-        $productFilterData = $this->productFilterDataFactory->create();
-        $productFilterData->minimalPrice = $frontendApiFilter['minimalPrice'] ?? null;
-        $productFilterData->maximalPrice = $frontendApiFilter['maximalPrice'] ?? null;
-        $productFilterData->parameters = $this->getParametersAndValuesByUuids($frontendApiFilter['parameters'] ?? []);
-        $productFilterData->inStock = $frontendApiFilter['onlyInStock'] ?? false;
-        $productFilterData->brands = [];
-        $productFilterData->flags = [];
-
-        if (isset($frontendApiFilter['brands'])) {
-            $productFilterData->brands = $this->brandFacade->getByUuids($frontendApiFilter['brands']);
-        }
-
-        if (isset($frontendApiFilter['flags'])) {
-            $productFilterData->flags = $this->flagFacade->getByUuids($frontendApiFilter['flags']);
-        }
-
-        return $productFilterData;
     }
 }
